+++ src/main/java/uk/ac/ed/inf/Menus.java
@@ -53,7 +53,7 @@ public class Menus
 				{
 					//iterate through the menus
 					for (int j = 0; j < items.menu.size();j++)
-			{
+					{
 						if ((items.menu.get(j).item).equals(foods.get(i)))
 						{
 							//adding the iterated items through the removing list
