+++ src/main/java/uk/ac/ed/inf/Shop.java
@@ -1,5 +1,6 @@
 package uk.ac.ed.inf;
 
+import java.util.HashMap;
 import java.util.List;
 
 // This is a class to set up gson parsing methods
@@ -7,11 +8,5 @@ public class Shop
 {
 	String name;
 	String location;
-
-	public List<MenuItems> menu;
-	public static class MenuItems
-	{
-		String item;
-		int pence;
-	}
+	public HashMap<String, Integer> menu;
 }
